% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{amariDistance}
\alias{amariDistance}
\title{Compute amari type distance between two matrices}
\usage{
amariDistance(matrix.A, matrix.B)
}
\arguments{
\item{matrix.A, matrix.B}{of the same dimensionality}
}
\value{
The amari type distance of matrix.A & matrix.B according
       to [Wu et. al, PNAS 2016]
}
\description{
Compute amari type distance between two matrices
}
\references{
\url{http://www.pnas.org/content/113/16/4290.long}
}

