\name{Bratwurst-package}
\alias{Bratwurst-package}
\alias{Bratwurst}
\docType{package}
\title{
\packageTitle{Bratwurst}
}
\description{
\packageDescription{Bratwurst}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Bratwurst}
\packageIndices{Bratwurst}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Bratwurst}

Maintainer: \packageMaintainer{Bratwurst}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
