% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{build_granges_object}
\alias{build_granges_object}
\title{Build GRanges object from signature list}
\usage{
build_granges_object(specific_signatures_list)
}
\arguments{
\item{specific_signatures_list}{}
}
\description{
Build GRanges object from signature list
}

