% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeAbsDelta}
\alias{computeAbsDelta}
\title{Compute delta between each column (signature) per row}
\usage{
computeAbsDelta(matrix)
}
\arguments{
\item{matrix}{}
}
\value{

}
\description{
Compute delta between each column (signature) per row
}

