% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeAbsDelta4OptK}
\alias{computeAbsDelta4OptK}
\title{Computes absolut delta per feature for each signature given optimal K}
\usage{
computeAbsDelta4OptK(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\description{
Computes absolut delta per feature for each signature given optimal K
}

