% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeAmariDistances}
\alias{computeAmariDistances}
\title{Compute Amari Distances from [Wu et. al, PNAS 2016]}
\usage{
computeAmariDistances(dec.matrix)
}
\arguments{
\item{nmf.exp}{}
}
\value{
The average Amari-type error for each k
}
\description{
Compute Amari Distances from [Wu et. al, PNAS 2016]
}
\references{
\url{http://www.pnas.org/content/113/16/4290.long}
}

