% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeCoefiVar}
\alias{computeCoefiVar}
\title{Compute the coeficient of variation per row in a matrix.}
\usage{
computeCoefiVar(matrix)
}
\arguments{
\item{matrix}{}
}
\value{

}
\description{
Compute the coeficient of variation per row in a matrix.
}

