% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeCopheneticCoeff}
\alias{computeCopheneticCoeff}
\title{Compute Cophenetic correlation coefficient, TO BE IMPROVED}
\usage{
computeCopheneticCoeff(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\description{
Compute Cophenetic correlation coefficient, TO BE IMPROVED
}

