% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeCpuFrobErrors}
\alias{computeCpuFrobErrors}
\title{Title}
\usage{
computeCpuFrobErrors(nmf.exp, cpu.nmf.list, k.min, k.max)
}
\arguments{
\item{outer.iter}{}
}
\value{

}
\description{
Title
}

