% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeEntropy}
\alias{computeEntropy}
\title{Compute 'shannon' entropy per region. 
High Entropy means highly specific for one signature.}
\usage{
computeEntropy(matrix)
}
\arguments{
\item{matrix}{}
}
\value{

}
\description{
Compute 'shannon' entropy per region. 
High Entropy means highly specific for one signature.
}

