% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeFeatureStats}
\alias{computeFeatureStats}
\title{Compute Signature Combinations for W-matrix given optimal K}
\usage{
computeFeatureStats(nmf.exp, var.thres = 0.25)
}
\arguments{
\item{var.thres}{}
}
\description{
Compute Signature Combinations for W-matrix given optimal K
}

