% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeFrobErrorStats}
\alias{computeFrobErrorStats}
\title{Compute basic statistics for Frobenius Errors}
\usage{
computeFrobErrorStats(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\description{
Compute basic statistics for Frobenius Errors
}

