% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeMeanDiff4SignatureCombs}
\alias{computeMeanDiff4SignatureCombs}
\title{Title}
\usage{
computeMeanDiff4SignatureCombs(W, i.regions)
}
\arguments{
\item{W}{}

\item{i.regions}{}
}
\description{
Title
}

