% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{computePeakStats4Regions}
\alias{computePeakStats4Regions}
\title{Title}
\usage{
computePeakStats4Regions(peaks, subregions, universe)
}
\arguments{
\item{peaks}{}

\item{subregions}{}

\item{universe}{}
}
\description{
Title
}

