% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computePValue4FrobError}
\alias{computePValue4FrobError}
\title{Compute p-value with t-test for running K}
\usage{
computePValue4FrobError(frobError.matrix)
}
\arguments{
\item{frobError.matrix}{}
}
\description{
Compute p-value with t-test for running K
}

