% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{computeRegionFC}
\alias{computeRegionFC}
\title{Title}
\usage{
computeRegionFC(peaks, subregions, universe)
}
\arguments{
\item{peaks}{}

\item{subregions}{}

\item{universe}{}
}
\description{
Title
}

