% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{computeRegionFC4GRlist}
\alias{computeRegionFC4GRlist}
\title{Title}
\usage{
computeRegionFC4GRlist(peak, subregion, universe)
}
\arguments{
\item{universe}{}
}
\value{

}
\description{
Title
}

