% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeSignatureCombs}
\alias{computeSignatureCombs}
\title{Compute Signature Combinations for W-matrix given optimal K}
\usage{
computeSignatureCombs(nmf.exp, k.opt, var.thres = 0.25)
}
\arguments{
\item{nmf.exp}{}

\item{var.thres}{}
}
\description{
Compute Signature Combinations for W-matrix given optimal K
}

