% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeSignatureSpecificFeatures}
\alias{computeSignatureSpecificFeatures}
\title{Compute signature specific features}
\usage{
computeSignatureSpecificFeatures(nmf.exp, rowDataId = 3)
}
\arguments{
\item{rowDataId}{The index of the rowData(nmf.exp) data.frame that should be used for 
feature extraction. In case rowData(nmf.exp)[,1] is a GRanges or a related object like
GenomicInteractions this parameter can be ignored}
}
\value{
nmf.exp with filles SignatureSpecificFeatures container
}
\description{
Compute signature specific features
}
\examples{

}
\references{

}

