% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeSilhoutteWidth}
\alias{computeSilhoutteWidth}
\title{Compute Alexandrov Criterion --> Silhoutte Width}
\usage{
computeSilhoutteWidth(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\value{

}
\description{
Compute Alexandrov Criterion --> Silhoutte Width
}

