% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{confusion_matrix_handmade}
\alias{confusion_matrix_handmade}
\title{Confusion matrix}
\usage{
confusion_matrix_handmade(temp_exposures_df, confusion_df, LCD_result_df,
  match_value)
}
\arguments{
\item{temp_exposures_df}{}

\item{confusion_df}{}

\item{LCD_result_df}{}

\item{match_value}{}
}
\description{
Confusion matrix
}

