% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{cosineDiss}
\alias{cosineDiss}
\title{Create distance matrix with cosine similarity}
\usage{
cosineDiss(in.matrix, in.dimension = 2)
}
\arguments{
\item{in.matrix}{}

\item{in.dimension}{}
}
\description{
Create distance matrix with cosine similarity
}

