% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{cosineDissMat}
\alias{cosineDissMat}
\title{Create distance matrix with cosine similarity with matrix operations}
\usage{
cosineDissMat(in.matrix, in.dimension = 2)
}
\arguments{
\item{in.dimension}{}
}
\description{
Create distance matrix with cosine similarity with matrix operations
}

