% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{extract_signature_specific_genes}
\alias{extract_signature_specific_genes}
\title{Extract genes specfic for each signature}
\usage{
extract_signature_specific_genes(genes_df, annotation_df,
  match_by = "external_gene_name")
}
\arguments{
\item{match_by}{}
}
\description{
Extract genes specfic for each signature
}

