% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{FeatureStats,nmfExperiment-method}
\alias{FeatureStats,nmfExperiment-method}
\title{Feature Statistics getter}
\usage{
\S4method{FeatureStats}{nmfExperiment}(x, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Feature Statistics getter
}

