% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{filter_features}
\alias{filter_features}
\title{Filter features}
\usage{
filter_features(k.row, out_int, specific_signatures, genes_df)
}
\arguments{
\item{genes_df}{}
}
\description{
Filter features which are specific for a single signature if min(size) == 1
is already set
}

