% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{filter_specific_features}
\alias{filter_specific_features}
\title{Filter features which are specific for a single signature}
\usage{
filter_specific_features(k.row, specific = TRUE)
}
\arguments{
\item{specific}{}
}
\description{
Filter features which are specific for a single signature
}

