% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{FrobError,nmfExperiment-method}
\alias{FrobError,nmfExperiment-method}
\title{Frobenius Error getter}
\usage{
\S4method{FrobError}{nmfExperiment}(x, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Frobenius Error getter
}

