% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_functions.R
\name{generate_classifier}
\alias{generate_classifier}
\title{Generate classifier}
\usage{
generate_classifier(matrix_W)
}
\arguments{
\item{matrix_W}{}
}
\value{
k.row, a list of lists (one list per row of the input matrix)
}
\description{
Details
}

