% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{generateRiverplot}
\alias{generateRiverplot}
\title{Generate a riverplot object that displays the similarities between
signatures at different factorization ranks}
\usage{
generateRiverplot(nmf.exp, edges.cutoff = 0, useH = FALSE, color = TRUE)
}
\arguments{
\item{edges.cutoff}{cutoff until which similarities are displayed}

\item{useH}{whether to relate signatures (FALSE) or exposures (TRUE)}

\item{color}{whether to colorize the nodes based on PCA of the signatures}
}
\value{
ret riverplot object
}
\description{
Generate a riverplot object that displays the similarities between
signatures at different factorization ranks
}
\examples{
plt <- generateRiverplot(nmf.exp, edges.cutoff = 0.2)
plot(plt, plot_area = 1, yscale = 0.6, nodewidth = 0.5)
}

