% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{getColorAnno}
\alias{getColorAnno}
\title{Computes color annotation for colData from NMF-experiment class}
\usage{
getColorAnno(col.data)
}
\arguments{
\item{col.data}{}
}
\value{

}
\description{
Computes color annotation for colData from NMF-experiment class
}

