% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{getColorMap}
\alias{getColorMap}
\title{Generate ColorMap for ComplexHeatmaps}
\usage{
getColorMap(matrix.list)
}
\arguments{
\item{matrix.list}{}
}
\value{

}
\description{
Generate ColorMap for ComplexHeatmaps
}

