% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{getCpuHMatrixList}
\alias{getCpuHMatrixList}
\title{Getter function for H-Matrix list from NMF-CPU list}
\usage{
getCpuHMatrixList(cpu.nmf.list, k.min, k.max)
}
\arguments{
\item{dec.matrix}{}
}
\value{

}
\description{
Getter function for H-Matrix list from NMF-CPU list
}

