% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{getCpuWMatrixList}
\alias{getCpuWMatrixList}
\title{Getter function for W-Matrix list from NMF-CPU list}
\usage{
getCpuWMatrixList(cpu.nmf.list, k.min, k.max)
}
\arguments{
\item{dec.matrix}{}
}
\value{

}
\description{
Getter function for W-Matrix list from NMF-CPU list
}

