% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{getFrobError}
\alias{getFrobError}
\title{Getter function for FrobError from NMF-GPU Object.}
\usage{
getFrobError(dec.matrix)
}
\arguments{
\item{dec.matrix}{}
}
\value{

}
\description{
Getter function for FrobError from NMF-GPU Object.
}

