% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{getHMatrixList}
\alias{getHMatrixList}
\title{Getter function for H-Matrix list from NMF-GPU list}
\usage{
getHMatrixList(dec.matrix)
}
\arguments{
\item{dec.matrix}{}
}
\value{

}
\description{
Getter function for H-Matrix list from NMF-GPU list
}

