% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{getSignatureNames}
\alias{getSignatureNames}
\title{Compute unsupervised signature names from colData and H-Matrix for given K}
\usage{
getSignatureNames(nmf.exp, k.opt)
}
\arguments{
\item{k.opt}{}
}
\description{
Compute unsupervised signature names from colData and H-Matrix for given K
}

