% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{getWMatrixList}
\alias{getWMatrixList}
\title{Getter function for W-Matrix list from NMF-GPU list}
\usage{
getWMatrixList(dec.matrix)
}
\arguments{
\item{dec.matrix}{}
}
\value{

}
\description{
Getter function for W-Matrix list from NMF-GPU list
}

