% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{HMatrix,nmfExperiment-method}
\alias{HMatrix,nmfExperiment-method}
\title{H-Matrix getter}
\usage{
\S4method{HMatrix}{nmfExperiment}(x, k = NULL, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
H-Matrix getter
}

