% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{HMatrixList,nmfExperiment-method}
\alias{HMatrixList,nmfExperiment-method}
\title{H-Matrix List getter}
\usage{
\S4method{HMatrixList}{nmfExperiment}(x, k = NULL, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
H-Matrix List getter
}

