% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{merge.nmf}
\alias{merge.nmf}
\title{Merge two objects of type nmfExperiment}
\usage{
merge.nmf(in_nmf1, in_nmf2, rerunStats = TRUE, verbose = FALSE)
}
\arguments{
\item{in_nmf1}{First object of type nmfExperiment}

\item{in_nmf2}{Second object of type nmfExperiment}

\item{rerunStats}{Boolean to indicate whether statistics should be computed on the merged
object}

\item{verbose}{}
}
\value{
The merged object of type nmfExperiment
}
\description{
Merge two objects of type nmfExperiment
}
\examples{
 NULL

}

