% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{class}
\name{nmfExperiment-class}
\alias{nmfExperiment}
\alias{nmfExperiment-class}
\title{NMF Experiment Class}
\value{

}
\description{
NMF Experiment Class
}
\section{Slots}{

\describe{
\item{\code{HMatrixList}}{list.}

\item{\code{HMatrix}}{list.}

\item{\code{WMatrixList}}{list.}

\item{\code{WMatrix}}{list.}

\item{\code{FrobError}}{DataFrame.}

\item{\code{OptKStats}}{DataFrame.}
}}

