% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{nmfExperimentFromMatrix}
\alias{nmfExperimentFromMatrix}
\title{Convert Matrix to SummarizedExperiment Object}
\usage{
nmfExperimentFromMatrix(matrix, row.anno = NULL, col.data = NULL)
}
\arguments{
\item{matrix}{}

\item{row.anno}{}

\item{col.data}{}
}
\value{

}
\description{
Convert Matrix to SummarizedExperiment Object
}
\examples{

}

