% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{nnls_sol}
\alias{nnls_sol}
\title{Find non-negative exposures of one matrix in another matrix}
\usage{
nnls_sol(B, A)
}
\arguments{
\item{B}{of A * X = B}

\item{A}{of A * X = B}
}
\value{
X of A * X = B
}
\description{
Find non-negative exposures of one matrix in another matrix
}

