% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{normalizeH}
\alias{normalizeH}
\title{Normalize the signatures matrix (H)}
\usage{
normalizeH(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\value{
A data structure of type nmfExperiment
}
\description{
After row normalization of the matrix H, the inverse factors are
mutiplied with the columns of W in order to keep the matrix product W*H
constant.
}
\examples{
 NULL

}

