% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{normalizeW}
\alias{normalizeW}
\title{Normalize the signatures matrix (W)}
\usage{
normalizeW(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\value{
A data structure of type nmfExperiment
}
\description{
After column normalization of the matrix W, the inverse factors are
mutiplied with the rows of H in order to keep the matrix product W*H
constant.
}
\examples{
 NULL

}

