% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{OptK,nmfExperiment-method}
\alias{OptK,nmfExperiment-method}
\title{Optimal K}
\usage{
\S4method{OptK}{nmfExperiment}(x, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Optimal K
}

