% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{OptKStats,nmfExperiment-method}
\alias{OptKStats,nmfExperiment-method}
\title{Optimal K Statistics getter}
\usage{
\S4method{OptKStats}{nmfExperiment}(x, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Optimal K Statistics getter
}

