% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{orderBinary}
\alias{orderBinary}
\title{Function to order binary data matrix}
\usage{
orderBinary(matrix)
}
\arguments{
\item{matrix}{}
}
\description{
Function to order binary data matrix
}

