% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{performRowKmeans}
\alias{performRowKmeans}
\title{Perform Kmeans on rows of a matrix to classify them into column (singature) combinations}
\usage{
performRowKmeans(matrix)
}
\arguments{
\item{matrix}{}
}
\value{

}
\description{
Perform Kmeans on rows of a matrix to classify them into column (singature) combinations
}

