% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotAmariError}
\alias{plotAmariError}
\title{Plots the average amari-type error over k
See [Wu et. al, PNAS 2016] for detais}
\usage{
plotAmariError(amari.error, keep.x = F)
}
\arguments{
\item{amari.error}{List of average amari error at different k}

\item{keep.x}{displays the x-axis}
}
\value{
Plot that displays the dependence of the amari error of k
        Smaller errors are better.
}
\description{
Plots the average amari-type error over k
See [Wu et. al, PNAS 2016] for detais
}

