% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotHeatmap4MatrixList}
\alias{plotHeatmap4MatrixList}
\title{Title}
\usage{
plotHeatmap4MatrixList(matrix.list, titles = NULL, trans = F,
  cluster_columns = F, show.row = F)
}
\arguments{
\item{matrix.list}{}

\item{titles}{}

\item{trans}{}

\item{cluster_columns}{}

\item{show.row}{}
}
\description{
Title
}

