% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotHMatrix}
\alias{plotHMatrix}
\title{Plot H-Matrix as heatmap for a given factorization rank K}
\usage{
plotHMatrix(nmf.exp, k = 2)
}
\arguments{
\item{nmf.exp}{}

\item{k}{}
}
\value{

}
\description{
Plot H-Matrix as heatmap for a given factorization rank K
}

