% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotKStats}
\alias{plotKStats}
\title{Universal ploting function for Optimal K Statistics}
\usage{
plotKStats(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\value{

}
\description{
Universal ploting function for Optimal K Statistics
}

