% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotRankedFrobErrors}
\alias{plotRankedFrobErrors}
\title{Ploting function for Rank vs. Frobenius Error}
\usage{
plotRankedFrobErrors(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\value{

}
\description{
Ploting function for Rank vs. Frobenius Error
}

