% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{readMatrix2SumExp}
\alias{readMatrix2SumExp}
\title{Reads Matrix & Annotation Files to SummarizedExperiment Object}
\usage{
readMatrix2SumExp(matrix.file, rowAnno.file = NULL, colData.file = NULL)
}
\arguments{
\item{matrix.file}{}

\item{rowAnno.file}{}

\item{colData.file}{}
}
\description{
Reads Matrix & Annotation Files to SummarizedExperiment Object
}

