% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{regularizeH}
\alias{regularizeH}
\title{Regularize the signatures matrix (H)}
\usage{
regularizeH(nmf.exp)
}
\arguments{
\item{nmf.exp}{}
}
\value{
A data structure of type nmfExperiment
}
\description{
After row regularization of the matrix H, the inverse factors are 
mutiplied with the columns of W in order to keep the matrix product W*H 
constant.
}
\examples{
 NULL

}

