% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{relabelRiverplot}
\alias{relabelRiverplot}
\title{Relabel and recolour a riverplot}
\usage{
relabelRiverplot(in_riverplot, in_list)
}
\arguments{
\item{in_list}{}
}
\description{
Relabel and recolour a riverplot
}

