% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{reorderEdges}
\alias{reorderEdges}
\title{Order the riverplot edges to prevent crossings from a single node}
\usage{
reorderEdges(nodes, edges)
}
\arguments{
\item{edges}{}
}
\value{
edges
}
\description{
Order the riverplot edges to prevent crossings from a single node
}

