% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{runNmfGpu}
\alias{runNmfGpu}
\title{Title}
\usage{
runNmfGpu(tmpMatrix.path, k.max = 2, outer.iter = 10, inner.iter = 10^4,
  out.dir = NULL)
}
\arguments{
\item{tmpMatrix.path}{}

\item{k.max}{}

\item{outer.iter}{}

\item{inner.iter}{}

\item{out.dir}{}
}
\description{
Title
}

