% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{runNmfGpuPyCuda}
\alias{runNmfGpuPyCuda}
\title{Title}
\usage{
runNmfGpuPyCuda(nmf.exp, k.min = 2, k.max = 2, outer.iter = 10,
  inner.iter = 10^4, conver.test.niter = 10,
  conver.test.stop.threshold = 40, out.dir = NULL,
  tmp.path = "/tmp/nmf_tmp", nmf.type = "N", w.sparsness = 0,
  h.sparsness = 0, gpu.id = 0, seed = FALSE,
  binary.file.transfer = FALSE, cpu = FALSE)
}
\arguments{
\item{out.dir}{}
}
\value{

}
\description{
Title
}

