% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{setFeatureStats,nmfExperiment-method}
\alias{setFeatureStats,nmfExperiment-method}
\title{Feature Statistics setter}
\usage{
\S4method{setFeatureStats}{nmfExperiment}(nmfExperiment, FeatureStats)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Feature Statistics setter
}

