% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{setFrobError,nmfExperiment-method}
\alias{setFrobError,nmfExperiment-method}
\title{Frobenius Error setter}
\usage{
\S4method{setFrobError}{nmfExperiment}(nmfExperiment, FrobError)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Frobenius Error setter
}

