% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{setHMatrixList,nmfExperiment-method}
\alias{setHMatrixList,nmfExperiment-method}
\title{H-Matrix List setter}
\usage{
\S4method{setHMatrixList}{nmfExperiment}(nmfExperiment, HMatrixList)
}
\arguments{
\item{nmfExperiment}{}
}
\value{

}
\description{
H-Matrix List setter
}

