% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{setOptK,nmfExperiment-method}
\alias{setOptK,nmfExperiment-method}
\title{Optimal K setter}
\usage{
\S4method{setOptK}{nmfExperiment}(nmfExperiment, OptK)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Optimal K setter
}

