% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{setOptKStats,nmfExperiment-method}
\alias{setOptKStats,nmfExperiment-method}
\title{Optimal K Statistics setter}
\usage{
\S4method{setOptKStats}{nmfExperiment}(nmfExperiment, OptKStats)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Optimal K Statistics setter
}

