% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{setSignatureSpecificFeatures,nmfExperiment-method}
\alias{setSignatureSpecificFeatures,nmfExperiment-method}
\title{Feature Statistics setter}
\usage{
\S4method{setSignatureSpecificFeatures}{nmfExperiment}(nmfExperiment,
  SignatureSpecificFeatures)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Feature Statistics setter
}

