% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{setWMatrixList,nmfExperiment-method}
\alias{setWMatrixList,nmfExperiment-method}
\title{W-Matrix setter}
\usage{
\S4method{setWMatrixList}{nmfExperiment}(nmfExperiment, WMatrixList)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
W-Matrix setter
}

