% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{SignatureSpecificFeatures,nmfExperiment-method}
\alias{SignatureSpecificFeatures,nmfExperiment-method}
\title{Signature specfific features getter}
\usage{
\S4method{SignatureSpecificFeatures}{nmfExperiment}(x, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
Signature specfific features getter
}

