% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizedExpNMF-class.R
\docType{class}
\name{summarizedExpNMF-class}
\alias{summarizedExpNMF}
\alias{summarizedExpNMF-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{HMatrixList}}{list.}

\item{\code{HMatrix}}{matrix.}

\item{\code{WMatrixList}}{list.}

\item{\code{WMatrix}}{matrix.}

\item{\code{FrobError}}{DataFrame.}
}}

