% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{WMatrix,nmfExperiment-method}
\alias{WMatrix,nmfExperiment-method}
\title{W-Matrix getter}
\usage{
\S4method{WMatrix}{nmfExperiment}(x, k = NULL, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
W-Matrix getter
}

