% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfExperiment-class.R
\docType{methods}
\name{WMatrixList,nmfExperiment-method}
\alias{WMatrixList,nmfExperiment-method}
\title{W-Matrix list getter}
\usage{
\S4method{WMatrixList}{nmfExperiment}(x, k = NULL, ...)
}
\arguments{
\item{nmfExperiment}{}
}
\description{
W-Matrix list getter
}

