% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{writeTmpMatrix}
\alias{writeTmpMatrix}
\title{Title}
\usage{
writeTmpMatrix(matrix, tmp.path = "/tmp/nmf_tmp", sep = " ")
}
\arguments{
\item{matrix}{}

\item{tmp.path}{}

\item{sep}{}
}
\value{

}
\description{
Title
}

