% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{yNodeCoords}
\alias{yNodeCoords}
\title{Order the riverplot nodes to minimize crossings}
\usage{
yNodeCoords(nodes, edges, rank_flag = FALSE, scale_factor = 1,
  start_coords = c(1, 2), edgeWeightColName = "rescaled",
  scale_fun = function(x) {     return(x) })
}
\arguments{
\item{sacle_fun}{}
}
\value{
node_ypos
}
\description{
Order the riverplot nodes to minimize crossings
}

