% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{computeEntropy4OptK}
\alias{computeEntropy4OptK}
\title{Computes entropy for each feature in optimal K W-matrix}
\usage{
computeEntropy4OptK(nmf.exp, opt.k)
}
\arguments{
\item{nmf.exp}{}

\item{opt.k}{}
}
\description{
Computes entropy for each feature in optimal K W-matrix
}

