% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{nmfExperimentFromFile}
\alias{nmfExperimentFromFile}
\title{Reads Matrix & Annotation Files to SummarizedExperiment Object}
\usage{
nmfExperimentFromFile(matrix.file, rowAnno.file = NULL, colData.file = NULL)
}
\arguments{
\item{colData.file}{}
}
\description{
Reads Matrix & Annotation Files to SummarizedExperiment Object
}

