% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotHeatmap4MatrixList}
\alias{plotHeatmap4MatrixList}
\title{Plot H or W-Matrix as heatmaps for a given range of K's}
\usage{
plotHeatmap4MatrixList(nmf.exp, H = T, W = F, titles = NULL, trans = F,
  cluster_columns = F, show.row = F)
}
\arguments{
\item{titles}{}

\item{trans}{}

\item{cluster_columns}{}

\item{show.row}{}

\item{matrix.list}{}
}
\value{

}
\description{
Plot H or W-Matrix as heatmaps for a given range of K's
}

