% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotKStats}
\alias{plotKStats}
\title{Universal ploting function for Optimal K Statistics}
\usage{
plotKStats(nmf.exp, plot.vars = c("FrobError", "cv", "sumSilWidth",
  "meanSilWidth", "copheneticCoeff", "meanAmariDist"), optK = NULL)
}
\arguments{
\item{plot.vars}{quality metrics to be displayed}

\item{optK}{optimal factorization rank, highlighted in the plot if non-NULL}
}
\value{

}
\description{
Universal ploting function for Optimal K Statistics
}

