% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{runNmfGpu}
\alias{runNmfGpu}
\title{Title}
\usage{
runNmfGpu(nmf.exp, k.min = 2, k.max = 2, outer.iter = 10,
  inner.iter = 10^4, conver.test.niter = 10,
  conver.test.stop.threshold = 40, out.dir = NULL,
  tmp.path = "/tmp/nmf_tmp")
}
\arguments{
\item{out.dir}{}
}
\value{

}
\description{
Title
}

