% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_functions.R
\name{runNmfGpuPyCuda}
\alias{runNmfGpuPyCuda}
\title{Title}
\usage{
runNmfGpuPyCuda(nmf.exp, k.min = 2, k.max = 2, outer.iter = 10,
  inner.iter = 10^4, conver.test.niter = 10,
  conver.test.stop.threshold = 40, out.dir = NULL,
  tmp.path = "/tmp/nmf_tmp", w.sparsness = 0, h.sparsness = 0)
}
\arguments{
\item{k.min}{}

\item{k.max}{}

\item{outer.iter}{}

\item{inner.iter}{}

\item{conver.test.niter}{}

\item{conver.test.stop.threshold}{}

\item{out.dir}{}
}
\value{

}
\description{
Title
}

